+++ RELEASE-NOTES.md
@@ -1,5 +1,10 @@
 Release notes
 -------------
+##### 6.0.0 (2022-05-14)
+ * Move getImage GraphQL query to getImages
+ * Add getImage GraphQL query
+ * Add image taken date and time as metadata
+ 
 ##### 5.4.1 (2022-05-06)
  * Update getImage GraphQL query signature
      - Move image type from ID to String